Gameplay Update (Server Side)
Too much done in this one commit, should of broken it up. But basic matchmaking, battle logic, teams, troops, units, read units from file

+++ AutoBattlerServer/src/main/java/com/team3/autobattlerserver/Network/PacketHandler.java
@@ -9,5 +9,5 @@ import org.json.JSONObject;
 public interface PacketHandler {
     // The collection of bytes which make up 
     // the user's request.
-    public void execute(int aId, JSONObject inputBuffer);
+    public void execute(long aId, JSONObject inputBuffer);
 }
