+++ AutoBattlerServer/src/main/java/com/team3/autobattlerserver/Network/Packets/Handle/GameStateChangePacket.java
@@ -25,7 +25,7 @@ public class GameStateChangePacket implements PacketHandler {
     
     
     @Override
-    public void execute(int aId, JSONObject inputBuffer) {
+    public void execute(long aId, JSONObject inputBuffer) {
         
         System.out.println("execute: Client Id: " + aId);
